We're using an intermediate model to explicitly manage the gene/transcript-history relationships.
From https://docs.djangoproject.com/en/2.0/topics/db/models/#many-to-many-relationships:
"Django allows you to specify the model that will be used to govern the many-to-many relationship.
You can then put extra fields on the intermediate model. The intermediate model is associated with
the ManyToManyField using the through argument to point to the model that will act as an intermediary"